+++ ChangeLog.md
@@ -9,7 +9,7 @@ and this project adheres to
 
 
 
-## [1.9.12] - 2021-02-xx
+## [1.9.12] - 2021-03-04
 
 ### Added